Trait rye_runtime::Spawner

pub trait Spawner {
    fn spawn(&mut self, fut: BoxFuture<'static, ()>) -> Result<()>;

    fn spawn_local(&mut self, fut: LocalBoxFuture<'static, ()>) -> Result<()>;

    fn spawn_blocking(
        &mut self, 
        f: Box<dyn FnOnce() + Send + 'static>
    ) -> Result<()>;
}

The value for spawning test cases.

Required methods

fn spawn(&mut self, fut: BoxFuture<'static, ()>) -> Result<()>

Spawn a task to execute a test case.

fn spawn_local(&mut self, fut: LocalBoxFuture<'static, ()>) -> Result<()>

Spawn a task to execute a test case onto the current thread.

fn spawn_blocking(
&mut self,
f: Box<dyn FnOnce() + Send + 'static>
) -> Result<()>

Spawn a task to execute a test case which may block the running thread.

Loading content...